Fix most memory leaks on error in thermodynamics
This method of calling _except macros to run free() commands on function
failure does not work well when memory is allocated to struct pointers.
This is because arrays like preco->recombination_table are allocated in
one function, and deallocated in another, and so would require calls to
free() throughout each subsequent function call which is cumbersome and
not particularly maintainable.

As a consequence, the thermodynamics module will still leak memory
during the allocation of:
 - preio->reionization_table
 - preco->recombination_table
 - pth->z_table
 - pth->thermodynamics_table
 - pth->d2thermodynamics_dz2_table