Fix more memory leaks on error in thermodynamics
Missed freeing preio->reionization_parameters in
thermodynamics_reionization() through _except method.

Add calls to free() on preio->reionization_parameters in same function
that were missing at the end of other reionisation parameterisations.
Note that preio->reionization_parameters is used in the
thermodynamics_reionization_sample() function, but this is only called
from within thermodynamics_reionization() and so no _except statements
need to be inserted, and the memory for the array should be deallocated
at the end.

For memory allocated to pointers that exist outside the scope of the
function, it may be best moving the _except calls to the scope at which
the pointer is defined at. For now, this will do but it is a lot of
work!