Fix memory leaks in thermodynamics.c and class.c
Memory allocated to generic_integrator() had not been deallocated in the
event of an errur during its execution.

Also copied the format for main() taken from the C++ wrappers, which
ensures all memory is deallocated properly depending on which module
CLASS fails in. This is still buggy, as there are no checks for where
the code fails in these modules (except for thermodynamics), and so a
call to <module>_free() may deallocate memory that has not yet been
allocated, giving a seg fault.