Stop CLASS if no external files have been read
If the 'read pbh splines' flag is set to false, and CLASS has not been
given a pointer to externally-read data, CLASS will crash with an
appropriate error.

The 'read pbh splines' flag should only be set to false when CLASS is
being called externally, and at some point in the parent program the
appropriate data files have been read so they can be passed to CLASS
instead of CLASS continually having to read these files in itself.