Enable external I/O of HyRec files
This should allow me to now run CLASS, but read the HyRec files in
beforehand, and pass the arrays in through thermodynamics_init() using
the new external_info structure.

Also made sure the HyRec on_the_spot flag is explicitly set using the
CLASS value, as this was not happening before and was giving
non-initialised conditional errors when putting it through Valgrind. In
this same vein, have also added freeing functions to the end of class.c
just to confirm that no memory leaks after a complete run of CLASS.