Add extern C to HyRec includes, logAlpha_tab stack
Wrapped all functions appearing in HyRec include files with extern 'C'
blocks, just to be sure that compiling with a C++ compiler knows it is dealing with C code.

Also put logAlpha_tab back onto the stack, in some regards. Do not need
the previous mallocation of just two since this is always constant. Can
now initialise the entries of this array (as it is on the stack) to the
two pointers that are set prior to this.