Add temperature tables to extern_info
extern_info now carries temperature table data as well as the tables'
steps. All pointer allocation should now make sense and work.

Also, 'correctly' deallocate preco->extern_hyrec. Previously, had a bare
free() which was not wrapped in a conditional statement. This was fine
because preco->extern_hyrec is set to NULL, and free(NULL) is OK
behaviour in the C standard.