Fix memory freeing errors in background & thermo
These were happening because unallocated arrays were being freed when a
module (usually in thermodynamics) failed. This has been fixed by
initialising arrays to NULL. free(NULL) is defined behaviour in C, and
does not lead to any memory corruption or errors. This is a quick fix
to get the ball rolling.