Fix warnings when compiling with -Wall
This also fixed a non-returning function in HyRec. Previously,
interpolate_rates() did not return _SUCCESS_, and so compilation on GNU
created a different result than on Intel. Both compilers behave
differently when a function returning an int doesn't actually do it.
With Intel, it seems to return an implicit zero, but on GNU whatever is
returned is evaluated to false and so breaks the program on the first
call to interpolate_rates().