Add array bound check on Dfminus_hist
In some cases, HyRec will overstep the bounds of the Dfminus_hist array
inside rec_HMLA_2photon_dxHIIdlna(). I believe this is to do with how
rec_dxHIIdlna() is called in history.c at smaller redshifts. As the
array index originates from the main integration loop in
rec_build_history(), the actual physics may prolong calls to certain
methods calling rec_dxHIIdlna(). This increases the iz index beyond the
end of the array, and as there was no check then this resulted in a seg
fault.