Add log normal PBH mass distribution to HyRec
The trapezoidal integration of injected energy, deposition fraction and
PBH mass distribution is yet to be validated. However, already know
injected energy and deposition fractions are calculated correctly from
the delta mass case. Have also validated the PBH mass distribution
generation externally before implementation.

Various checks occur on the distribution that are worth outlining here:
- If the left of the distribution falls sufficiently out of the PBH mass
range, the total distribution is renormalised such that the truncated
distribution integrates to unity. This is done so as to not lose extra
injected energy from low-mass, high-temperature PBHs which the
recombination history will be sensitive to. The actual cutoff is defined
by the _SIGMA_TRUNCATE_RANGE_ macro in hyrec/include/energy_injection.h.
- The right side truncation is ignored, as this end corresponds to high-
mass, low-temperature PBHs which inject comparatively less and as such
can be effectively ignored (to what extent?).
- Full integration occurs +- 4*sigma of the distribution, and can be
similarly changed with the _SIGMA_INTEGRAL_RANGE_ macro in the same
file.