Fix seg fault in perturb_timesampling_for_sources()
CLASS seg faulted if any test in perturb_timesampling_for_sources()
failed. This was because the main ppr->sources array of array of arrays
had not yet been fully allocated, so upon failure of
perturb_timesampling_for_sources(), CLASS would attempt to call
perturb_free(), which of course would fail on the non-allocated memory.

This was fixed by defining a flag which is set to _TRUE_ when the full
array is allocated, and is _FALSE_ otherwise. perturb_free() then checks
this flag.

The best solution to this problem (and many others) is to initialise all
of these dynamic pointers to NULL. Because free(NULL) is perfectly well-
defined behaviour in C (it does nothing bad with memory), this will
avoid all future seg faulting issues based on non-allocated memory.