Fix overestimation in HyRec due to implicit unity
When computing Alpha_eq in interpolate_rates(), the fact that the Tm/Tr
table ends at Tm/Tr = 1.0 is used implicitly to set Tm = Tr. The new
table ends at Tm/Tr = 200.0, and so Alpha_eq is calculated incorrectly
(instead of Tm = Tr, Tm = 200Tr!). The offending indices have been
renamed and redefined for the new table to take the value closest to
Tm/Tr = 1.0. Look for TM_TR_UNITY_IND in io_params.h.

A slight issue is that the new table does not have a value in it that
is exactly equal to unity. The closest value is 1.00072, which seems to
create issues for large energy injections.