Fix premature memory freeing in input.c
input_try_unknown_parameters() does its job by running CLASS modules in
ascending order, up to a module defined by
pfzw->required_computation_stage. On completion, the rest of the code
deallocates memory according to the same required computation stage
value. However, if CLASS fails in any modules leading up to the final
one necessary for the calculation, then it will begin by attempting to
free memory that has not even been initialised, let alone allocated,
leading to a seg fault.

To fix this, a second variable "stage_reached" was created such that
this keeps track of the actual stage that was reached during
computation, meaning only memory that was properly initialised and
allocated will be freed in the event of a failure earlier in the
pipeline.