Change pbh_spline_files_root to specify full path
This was required in cases where the directory for the PBH spline files
did not exist in the current directory. Plus, I feel like it makes more
sense this way if for whatever reason someone has their PBH files in a
different place, or are running a program at a non-standard location.

This parameter now must specify the _full_ path to the PBH splines, as
well as their root name which by default is "pbh_bspline_".